Fix issues in tests
