+++ fe2-android/app/src/test/framework/common/java/com/github/dedis/popstellar/testutils/MatcherUtils.java
@@ -1,0 +1,28 @@
+package com.github.dedis.popstellar.testutils;
+
+import android.view.*;
+
+import org.hamcrest.*;
+
+public class MatcherUtils {
+
+  public static Matcher<View> childAtPosition(
+      final Matcher<View> parentMatcher, final int position) {
+
+    return new TypeSafeMatcher<View>() {
+      @Override
+      public void describeTo(Description description) {
+        description.appendText("Child at position " + position + " in parent ");
+        parentMatcher.describeTo(description);
+      }
+
+      @Override
+      public boolean matchesSafely(View view) {
+        ViewParent parent = view.getParent();
+        return parent instanceof ViewGroup
+            && parentMatcher.matches(parent)
+            && view.equals(((ViewGroup) parent).getChildAt(position));
+      }
+    };
+  }
+}
