+++ fe2-android/app/src/test/framework/emulator/java/com/github/dedis/popstellar/testutils/UITestUtils.java
@@ -2,13 +2,10 @@ package com.github.dedis.popstellar.testutils;
 
 import android.app.Dialog;
 import android.content.DialogInterface;
-import android.view.*;
 import android.widget.Button;
 
 import androidx.annotation.StringRes;
 import androidx.appcompat.app.AlertDialog;
-
-import org.hamcrest.*;
 
 /** This class holds utility functions when retrieving particular elements of a view in a test */
 public class UITestUtils {
@@ -70,25 +67,5 @@ public class UITestUtils {
     } else {
       throw new AssertionError("The dialog does not have a positive button");
     }
-  }
-
-  public static Matcher<View> childAtPosition(
-      final Matcher<View> parentMatcher, final int position) {
-
-    return new TypeSafeMatcher<View>() {
-      @Override
-      public void describeTo(Description description) {
-        description.appendText("Child at position " + position + " in parent ");
-        parentMatcher.describeTo(description);
-      }
-
-      @Override
-      public boolean matchesSafely(View view) {
-        ViewParent parent = view.getParent();
-        return parent instanceof ViewGroup
-            && parentMatcher.matches(parent)
-            && view.equals(((ViewGroup) parent).getChildAt(position));
-      }
-    };
   }
 }
