+++ fe2-android/app/src/test/ui/robolectric/com/github/dedis/popstellar/ui/pages/detail/event/election/ElectionSetupPageObject.java
@@ -6,7 +6,7 @@ import com.github.dedis.popstellar.R;
 
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.matcher.ViewMatchers.*;
-import static com.github.dedis.popstellar.testutils.UITestUtils.childAtPosition;
+import static com.github.dedis.popstellar.testutils.MatcherUtils.childAtPosition;
 import static org.hamcrest.Matchers.allOf;
 
 public class ElectionSetupPageObject {
